requirements.txt
certifi==2017.11.5
chardet==3.0.4
decorator==4.1.2
defusedxml==0.5.0
Django==2.0.1
django-allauth==0.34.0
django-cors-headers==2.1.0
django-db-constraints==0.3.0
django-rest-auth==0.9.2
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
idna==2.6
intriniorealtime==2.0.0
numpy==1.14.0
oauthlib==2.0.6
pandas==0.22.0
protobuf==3.7.1
psycopg2==2.7.3.2
pyiqfeed==0.9
PyJWT==1.5.3
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
websocket-client==0.46.0
yahoo-earnings-calendar==0.3.1
